requirements.txt
django
south
django-debug-toolbar
lxml
pygments
docutils
django-mptt
BeautifulSoup
django-urlauth
